Added scan for metadata on uploaded files
Javascript is used to scan the uploaded files locally for metadata. At the moment, only PDF files are supported. If metadata is found, the user is asked to confirm that it is acceptable. If the file cannot be scanned, the user is asked to confirm that they are sure about the file's metadata.